+++ cla/tcpclv4/0doc.go
@@ -2,9 +2,9 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-// Package tcpcl provides a library for the Delay-Tolerant Networking TCP Convergence Layer Protocol Version 4,
+// Package tcpclv4 provides a library for the Delay-Tolerant Networking TCP Convergence Layer Protocol Version 4,
 // draft-ietf-dtn-tcpclv4-14.
 //
 // A new TCPCL server can be started by the TCPListener, which provides multiple connection to its Clients. To reach
 // a remote server, a new Client connection can be dialed.
-package tcpcl
+package tcpclv4