+++ cla/tcpclv4/internal/stages/contact.go
@@ -7,7 +7,7 @@ package stages
 import (
 	"fmt"
 
-	"github.com/dtn7/dtn7-go/cla/tcpcl/internal/msgs"
+	"github.com/dtn7/dtn7-go/cla/tcpclv4/internal/msgs"
 )
 
 // ContactStage models the initial ContactHeader exchange.