+++ cla/tcpclv4/internal/stages/sess_established_test.go
@@ -12,7 +12,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/dtn7/dtn7-go/cla/tcpcl/internal/msgs"
+	"github.com/dtn7/dtn7-go/cla/tcpclv4/internal/msgs"
 )
 
 func TestSessEstablishedStageKeepalivePingPong(t *testing.T) {