+++ cla/tcpclv4/internal/stages/sess_init_test.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	"github.com/dtn7/dtn7-go/bundle"
-	"github.com/dtn7/dtn7-go/cla/tcpcl/internal/msgs"
+	"github.com/dtn7/dtn7-go/cla/tcpclv4/internal/msgs"
 )
 
 func TestSessInitStage(t *testing.T) {