+++ cla/tcpclv4/internal/stages/stage.go
@@ -8,7 +8,7 @@ import (
 	"errors"
 
 	"github.com/dtn7/dtn7-go/bundle"
-	"github.com/dtn7/dtn7-go/cla/tcpcl/internal/msgs"
+	"github.com/dtn7/dtn7-go/cla/tcpclv4/internal/msgs"
 )
 
 // Configuration for stages.