+++ cla/tcpclv4/internal/utils/transfer_manager.go
@@ -13,7 +13,7 @@ import (
 	"time"
 
 	"github.com/dtn7/dtn7-go/bundle"
-	"github.com/dtn7/dtn7-go/cla/tcpcl/internal/msgs"
+	"github.com/dtn7/dtn7-go/cla/tcpclv4/internal/msgs"
 )
 
 // TransferManager transfers Bundles bidirectionally.