+++ cla/tcpclv4/internal/utils/transfer_out.go
@@ -10,7 +10,7 @@ import (
 	"io"
 
 	"github.com/dtn7/dtn7-go/bundle"
-	"github.com/dtn7/dtn7-go/cla/tcpcl/internal/msgs"
+	"github.com/dtn7/dtn7-go/cla/tcpclv4/internal/msgs"
 )
 
 // OutgoingTransfer represents an outgoing Bundle Transfer for the TCPCL.