+++ cla/tcpclv4/network_test.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-package tcpcl
+package tcpclv4
 
 import (
 	"fmt"